#1890: extract coordinator/cos_leases.rs (CoS runtime-map plumbing)
Third pure code-motion extraction for the #1890 split: move the CoS
runtime-map block into coordinator/cos_leases.rs — the three
Coordinator methods (refresh_cos_owner_worker_map_from_identities,
refresh_cos_owner_worker_map_from_binding_statuses,
refresh_cos_runtime_maps) and the free-function builders/matchers they
delegate to (owner-by-queue assignment, active-shard counts,
root/queue lease + exact-backlog + vtime-floor construction with the
Arc-reuse discipline, the Arc::ptr_eq match predicates, and the #710
cross-worker status aggregation consumed by status.rs).

build_mirror_target_map stays in mod.rs: it is port-mirroring
plumbing, not CoS — it merely sat inside the moved byte range.

Visibility deltas, each forced by the move and staying inside the
coordinator subtree (pub(super)):
- the three refresh methods (callers: snapshot_refresh.rs,
  refresh_bindings.rs, reconcile/bringup.rs, tests.rs);
- the builders with out-of-file consumers (bringup.rs fixtures paths,
  tests.rs fixtures).
Match predicates and the two builders with no out-of-file callers
(build_cos_owner_live_by_queue,
build_shared_cos_exact_backlogs_reusing_existing) stay private.
mod.rs re-imports the free functions so pre-split coordinator-scope
references (status.rs / tests.rs via `use super::*`, bringup.rs via
`super::super::`) resolve unchanged; the re-imports whose only
consumers are tests.rs fixtures are #[cfg(test)]-gated to keep the
release warning set byte-identical to baseline. Two moved signatures
(refresh_cos_owner_worker_map_from_binding_statuses,
unique_interface_owner_worker_id) are re-wrapped because the
pub(super) token pushed them past 100 columns — no semantic change.

Validation: cargo build --release clean, warning set byte-identical to
the origin/master baseline; cargo test --release --no-run compiles all
test targets. Full test gates run at the end of the series.
